pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devflow-ai"
version = "0.2.0"
description = "Spec-driven build orchestrator for AI coding agents — plan-first workflow, quality gate, multi-provider, cost tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Justine Raze" }]
keywords = ["cli", "ai", "claude", "development", "workflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.scripts]
devflow = "devflow.cli:app"

[project.urls]
Homepage = "https://github.com/JustineRaze/devflow-ai"
Repository = "https://github.com/JustineRaze/devflow-ai"

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
files = ["src"]
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.conftest", "tests.contract.*"]
strict = true

[[tool.mypy.overrides]]
module = ["tests.unit.*", "tests.e2e.*", "tests.smoke.*"]
disallow_untyped_defs = false
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/e2e", "tests/smoke", "tests/contract"]
addopts = "-m 'not smoke' --cov=src/devflow --cov-report=term-missing --cov-fail-under=80"
norecursedirs = ["tests/e2e/fixtures", "tests/smoke/fixtures"]
markers = [
    "e2e: orchestration tests — real git + gate, mocked claude (fast, no API cost)",
    "smoke: full pipeline with real claude -p calls — run manually before releases",
    "contract: backend protocol conformance tests — run with -m contract",
]

[tool.coverage.run]
branch = false
source = ["src/devflow"]
omit = [
    "src/devflow/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.hatch.build.targets.wheel]
packages = ["src/devflow"]

[tool.hatch.build.targets.wheel.force-include]
"assets" = "assets"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=5.0",
    "ruff>=0.15.10",
    "mypy>=1.13",
    "types-pyyaml>=6.0",
    "import-linter>=2.0",
]

[tool.importlinter]
root_packages = ["devflow"]
include_external_packages = false

[[tool.importlinter.contracts]]
name = "Core layering"
type = "layers"
layers = [
    "devflow.cli",
    "devflow.orchestration",
    "devflow.integrations",
    "devflow.core",
]

[[tool.importlinter.contracts]]
name = "UI only imports core"
type = "forbidden"
source_modules = ["devflow.ui"]
forbidden_modules = [
    "devflow.orchestration",
    "devflow.integrations",
    "devflow.setup",
]

[[tool.importlinter.contracts]]
name = "Setup only imports core"
type = "forbidden"
source_modules = ["devflow.setup"]
forbidden_modules = [
    "devflow.orchestration",
    "devflow.integrations",
    "devflow.ui",
]
